Add files via upload
To reverse a string in Python, we will use slicing. Specifically, we will use [::-1]. Here we have not provided the start and end, therefore slicing would happen from start to end of the string. -1 would make the slicing happen in reverse, i.e., from end to start. And str[::-1] would result in a reversed string.

** Reverse String using For Loop** we will reverse a string using Python For Loop. With for loop, we iterate over characters in string and append each character at the front of a new string(initially empty). By the end of the for loop, we should have a reversed string.